Moving sleep from window to thread namespace
