+++ magnolia/src/threads/thread.cpp
@@ -1,0 +1,30 @@
+#include "threads/thread.hpp"
+
+#include <chrono>
+#include <ratio>
+#include <thread>
+
+namespace mag
+{
+    namespace thread
+    {
+        b8 initialize_job_system(const u32 max_number_of_threads);
+        void shutdown_job_system();
+
+        b8 initialize()
+        {
+            const u32 max_number_of_threads = std::thread::hardware_concurrency();
+            const b8 initialized = initialize_job_system(max_number_of_threads);
+
+            return initialized;
+        }
+
+        void shutdown() { shutdown_job_system(); }
+
+        void sleep(const f64 ms)
+        {
+            std::chrono::duration<f64, std::milli> duration(ms);
+            std::this_thread::sleep_for(duration);
+        }
+    };  // namespace thread
+};      // namespace mag
